+++ src/main/java/project/personal/lhinfo/dao/TypeDao.java
@@ -1,34 +1,18 @@
 package project.personal.lhinfo.dao;
 
-import org.mybatis.spring.SqlSessionTemplate;
-import org.springframework.stereotype.Repository;
 import project.personal.lhinfo.entity.Location;
 import project.personal.lhinfo.entity.NoticeStatusType;
 import project.personal.lhinfo.entity.NoticeType;
 import project.personal.lhinfo.entity.SupplyType;
 
-import javax.annotation.Resource;
 import java.util.List;
 
-@Repository("typeDao")
-public class TypeDao {
+public interface TypeDao {
+    List<Location> locationList();
 
-    @Resource(name = "sqlSessionTemplate")
-    private SqlSessionTemplate session;
+    List<SupplyType> supplyTypeList();
 
-    public List<Location> locationList() {
-        return session.selectList("lhinfoDB.selectLocations");
-    }
+    List<NoticeType> noticeTypeList();
 
-    public List<SupplyType> supplyTypeList() {
-        return session.selectList("lhinfoDB.selectSupplyTypes");
-    }
-
-    public List<NoticeType> noticeTypeList() {
-        return session.selectList("lhinfoDB.selectNoticeTypes");
-    }
-
-    public List<NoticeStatusType> noticeStatusTypeList() {
-        return session.selectList("lhinfoDB.selectNoticeStatusTypes");
-    }
+    List<NoticeStatusType> noticeStatusTypeList();
 }
